+++ database/migrations/2025_04_29_154537_create_tasks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
       $table->foreignIdFor(Column::class)->constrained()->cascadeOnDelete();
       $table->string("title");
       $table->text("description")->nullable();
-      $table->unsignedInteger('position');
+      $table->double('position');
       $table->timestamps();
     });
   }
